refactor: Sections 일급콜렉션 추가
